// lib/api/generated/model/documentRead.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Syncnesto API
 * OpenAPI spec version: 0.1.0
 */

/**
 * 現在の本文と更新者。
 */
export interface DocumentRead {
  id: number;
  project_id: number;
  title: string;
  body: string;
  version: number;
  created_by: number;
  updated_by: number;
  created_at: string;
  updated_at: string;
}